HZN-1073: Rename all the things
