+++ src/datasources/fault-ds/Mapping.js
@@ -1,0 +1,17 @@
+import {ComparatorMapping} from './mapping/ComparatorMapping';
+import {OperatorMapping} from './mapping/OperatorMapping';
+import {AttributeMapping} from './mapping/AttributeMapping';
+import {FilterMapping} from './mapping/FilterMapping';
+import {RestrictionMapping} from './mapping/RestrictionMapping';
+import {ClauseMapping} from './mapping/ClauseMapping';
+import {ValueMapping} from './mapping/ValueMapping';
+
+export const Mapping = Object.freeze({
+    ComparatorMapping,
+    OperatorMapping,
+    AttributeMapping,
+    FilterMapping,
+    RestrictionMapping,
+    ClauseMapping,
+    ValueMapping
+});
