--- src/datasources/fault-ds/mapping.js
@@ -1,15 +1,0 @@
-import {ComparatorMapping} from './mapping/comparator';
-import {OperatorMapping} from './mapping/operator';
-import {AttributeMapping} from './mapping/attribute';
-import {FilterMapping} from './mapping/filter';
-import {RestrictionMapping} from './mapping/restriction';
-import {ClauseMapping} from './mapping/clause';
-
-export const Mapping = Object.freeze({
-    ComparatorMapping,
-    OperatorMapping,
-    AttributeMapping,
-    FilterMapping,
-    RestrictionMapping,
-    ClauseMapping
-});
