+++ src/datasources/fault-ds/mapping/OperatorMapping.js
@@ -1,5 +1,4 @@
 import _ from 'lodash';
-import {UI} from '../ui';
 import {API} from '../../../opennms'
 
 export class OperatorMapping {
